Fixing integer overflow issue in RTS metrics (#6539)

+++ test/WebJobs.Script.Tests.Shared/TestScaleMonitor.cs
@@ -61,7 +61,7 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
 
     public class TestScaleMetrics1 : ScaleMetrics
     {
-        public int Count { get; set; }
+        public long Count { get; set; }
     }
 
     public class TestScaleMonitor2 : TestScaleMonitor<TestScaleMetrics2>
